+++ Source/SFTargetSearchTests/Private/Mocks/TargetSearch_MockInstigatorObject.h
@@ -1,0 +1,16 @@
+﻿// Copyright Strayfarer & Contributors. Released under the MIT license.
+
+#pragma once
+
+#include "CoreMinimal.h"
+#include "GameFramework/Actor.h"
+#include "TargetSearch_MockInstigatorObject.generated.h"
+
+/**
+ * Mock instigator object used for testing.
+ */
+UCLASS(NotBlueprintType, Hidden)
+class UTargetSearch_MockInstigator : public UObject
+{
+	GENERATED_BODY()
+};