+++ CHANGELOG.md
@@ -6,7 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+### Added
+- Added unit tests
+### Changed
+- Made query result struct members read-only
+### Fixed
+- Removed ComponentPicker from module dependencies
 
 ## [1.0.0] - 2026-05-18